Add new tool ni2out
Add new tool ni2out that is currently a copy of ni2http.
It accepts SID from command line and it uses stdout to print the output.
It requires no configuration file like ni2http, in order to be used in command line.

Additionally add ni2out and install targets in Makefile in order to use it
from OpenEmbedded.